xserver xkeyboard-config fontconfig mesa git update Jan 17 2018
xserver          commit 32b42627217917337553121191e9bc3b423b8d99
xkeyboard-config commit f734a19420ab6d37dee51c6ace7d6b6aeb85f967
fontconfig       commit 73cc842d1dd866e4a6fda4aa422cb4a9c7a9832f
mesa             commit f631a6ae8c951fc448618c1770f95ad93cac812b

+++ mesalib/src/gallium/auxiliary/hud/hud_sensors_temp.c
@@ -165,19 +165,19 @@ query_sti_load(struct hud_graph *gr)
 
          switch (sti->mode) {
          case SENSORS_TEMP_CURRENT:
-            hud_graph_add_value(gr, (uint64_t) sti->current);
+            hud_graph_add_value(gr, sti->current);
             break;
          case SENSORS_TEMP_CRITICAL:
-            hud_graph_add_value(gr, (uint64_t) sti->critical);
+            hud_graph_add_value(gr, sti->critical);
             break;
          case SENSORS_VOLTAGE_CURRENT:
-            hud_graph_add_value(gr, (uint64_t)(sti->current * 1000));
+            hud_graph_add_value(gr, sti->current * 1000);
             break;
          case SENSORS_CURRENT_CURRENT:
-            hud_graph_add_value(gr, (uint64_t) sti->current);
+            hud_graph_add_value(gr, sti->current);
             break;
          case SENSORS_POWER_CURRENT:
-            hud_graph_add_value(gr, (uint64_t) sti->current);
+            hud_graph_add_value(gr, sti->current);
             break;
          }
